refactor: modularize generators and actions
- break plopfile generator and action logic in generators and actions
- loading them asynchronously
- modularity reduces redundancy and increases scalability
